requirements.txt
flask
openai
python-dotenv
Pillow
opencv-python-headless
numpy
ultralytics
textwrap3
pdf2image
PyMuPDF
tqdm
